+++ src/index.js
@@ -24,13 +24,16 @@ export default class ContainerDimensions extends Component {
 
     componentDidMount() {
         this.parentNode = ReactDOM.findDOMNode(this).parentNode
-        this.elementResizeDetector = elementResizeDetectorMaker({ strategy: 'scroll' })
+        this.elementResizeDetector = elementResizeDetectorMaker({
+            strategy: 'scroll',
+            callOnAdd: false
+        })
         this.elementResizeDetector.listenTo(this.parentNode, this.onResize)
         this.onResize()
     }
 
     componentWillUnmount() {
-        this.elementResizeDetector.removeListener(this.parentNode, this.onResize)
+        this.elementResizeDetector.uninstall(this.parentNode)
     }
 
     onResize() {