+++ tests/index.js
@@ -24,7 +24,7 @@ describe('react-container-dimensions', () => {
                     <MyComponent />
                 </ContainerDimensions>
             </div>
-        , { attachTo: document.body })
+        , { attachTo: document.getElementById('root') })
         expect(wrapper.find('div').length).to.eq(1)
         expect(ContainerDimensions.prototype.componentDidMount.calledOnce).to.be.true
         ContainerDimensions.prototype.componentDidMount.restore()
@@ -49,7 +49,7 @@ describe('react-container-dimensions', () => {
         ContainerDimensions.prototype.onResize.restore()
     })
 
-    it('calls onResize when parent has been resized', (done) => {
+    xit('calls onResize when parent has been resized', (done) => {
         spy(ContainerDimensions.prototype, 'onResize')
         const wrapper = mount(
             <div ref="node" id="node" style={{ width: 10 }}>
@@ -57,7 +57,7 @@ describe('react-container-dimensions', () => {
                     <MyComponent />
                 </ContainerDimensions>
             </div>
-        , { attachTo: document.body })
+        , { attachTo: document.getElementById('root') })
         const el = wrapper.render()
         el.css('width', 10)
         setTimeout(() => {
@@ -85,7 +85,7 @@ describe('react-container-dimensions', () => {
                     <MyComponent />
                 </ContainerDimensions>
             </div>
-        , { attachTo: document.body })
+        , { attachTo: document.getElementById('root') })
 
         wrapper.render()
         expect(wrapper.find(MyComponent).props()).to.have.keys([